refactoring webcli CodeEditor component
