+++ jweb-console-webcli/src/components/CodeEditor/constants.ts
@@ -1,0 +1,1 @@
+export const NO_ENGINE = "no-engine";
